feat(m): create class (schema) for tables in db
`SQLAlchemy` uses classes to declare the schema for the tables in the
database.

The schema file _dbschema.py_ declares classes for and one-to-many
relationships with:
    * `User`: `users` table -> `purchases` table,
    * `Purchase`: `purchases` table -> `payments` table and
    * `Payment`: `payments` table.

That is, one user can have many purchases, which, in turn, can have many
payments themselves. However, a payment can only have one purchase,
which, in turn, can only have one user.