feat(c:payment): implement payment
This should allow the user to add a new payment to an existing "Pending"
purchase.

When adding the payment, it should update the debt the purchase's
current debt as well as the user's. If the payment sets the purchase's
debt to `0` (zero), it should also update its status to "Cleared".