feat(v:payments): show all payments for the current user
Due to implementation limitations, the column creation for the table
cannot be done entirely programmatically. This, since I'm passing the
header names to a jinja template as their field name, e.g. "amount".

Doing things like this allows me to access their value ala
`payment.amount` and keep the column title as "Amount"; the view is
human readable and the controller is organised and simple.

However, for the payments I want to have information about the purchase
and not only the payment itself.

With the current implementation, this would require to pass the header
name as "purchase.item", which, in turn, would make the column title
into "Purchase.Item".